requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==23.1.0
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.32.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
defusedxml==0.7.1
dj-database-url==1.3.0
Django==4.2
django-allauth==0.54.0
django-allauth-2fa==0.10.0
django-extensions==3.2.1
django-leaflet==0.29.0
django-otp==1.2.1
dnspython==2.3.0
eventlet==0.33.3
filelock==3.12.2
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
parsel==1.8.1
pika==1.3.1
Pillow==9.5.0
prompt-toolkit==3.0.38
Protego==0.2.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
PyDispatcher==2.0.7
PyJWT==2.6.0
pyOpenSSL==23.1.1
pypng==0.20220715.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
qrcode==7.4.2
queuelib==1.6.2
rabbitmq==0.2.0
rave-python==1.2.16
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.9.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
w3lib==2.1.1
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.0
whitenoise==6.4.0
zope.interface==6.0
